Refactor LessonsListScreen utilities
